Update paper.md
In first paragraph of Statement of need, changed "available in" to "from" to reduce the use of "available" in the sentence and paragraph.

Additional suggestion: should we include a clear statement in the Summary section (maybe 2nd paragraph) that the package includes the links to relevant data (no need to search for the data)? My thought on this is that many partners ask for a "centralized" data acquisition option.